501 darts simulator mimicking Luke Littler
Learning project v1.00; 

This a very basic, but functional not without bugs AI simulator based on the stats of the 2024 Luke Littler. However this is my very first project with no Python experience. 
There are current bugs with the coding, such as improper values can be entered, and the game can enter an endless state or if bust rules are not enforced.

As well there might be additional bugs I am unaware of but this is my first project with no Python training or experience.  

Bugs encountered;
Luke's average ranges & checkout ranges being shown to the player; 
This was corrected by removing print function. 

Coin toss was not giving the elected player the throw;
This was corrected by sorting the winner of the coin toss in the thrower variable. 

The turns wouldn't swap after input value so added the current_player function variable to switch. 

Lastly, the AI would check out random variables post-unrealistic checkout scores such as 249 so added in parameters over viable checkouts.